minor change
git-svn-id: https://svn.eu.apache.org/repos/asf/synapse/trunk@745419 13f79535-47bb-0310-9956-ffa450edef68

+++ java/modules/core/src/main/java/org/apache/synapse/config/xml/endpoints/EndpointFactory.java
@@ -302,7 +302,7 @@ public abstract class EndpointFactory implements XMLToObjectMapper {
             log.warn("Configuration uses deprecated style for endpoint 'suspendDurationOnFailure'");
             try {
                 definition.setInitialSuspendDuration(
-                    1000 * Integer.parseInt(suspendDurationOnFailure.getText().trim()));
+                        1000 * Long.parseLong(suspendDurationOnFailure.getText().trim()));
                 definition.setSuspendProgressionFactor((float) 1.0);
             } catch (NumberFormatException e) {
                 handleException("The initial suspend duration should be specified " +
